requirements.txt
playwright
beautifulsoup4
streamlit
google-genai
python-dotenv
streamlit-js-eval